+++ Paramount.Betterclassifieds.Business/Events/Booking/EventBookingContext.cs
@@ -38,6 +38,8 @@
             Purchaser = null;
             EventInvitationId = null;
             OrderRequestId = null;
+            AppliedPromoCode = null;
+            EventUrl = null;
         }
     }
 }