+++ ReleaseNotes.md
@@ -2,6 +2,11 @@ Here's a list of the latest changes:
 
 **3.28.2**
 
+- Fix - The seat was not booked properly when 100 percent discount was applied.
+- Fix - The calculation of the 100 percent discount was still adding cents!
+
+**3.28.2**
+
 - Fix - The account login wasn't handling the errors from the server properly
 
 **3.28.1**